5476 - Add inner hook
